absensi

+++ app/Http/Controllers/Admin/Referensi/TingkatanKelasController.php
@@ -7,27 +7,29 @@ use Illuminate\Http\Request;
 use App\Models\TingkatanKelas;
 use Yajra\DataTables\DataTables;
 use App\Http\Controllers\Controller;
+use App\User;
 
 class TingkatanKelasController extends Controller
 {
-    public function index(Request $request) {
+    public function index(Request $request)
+    {
         if ($request->ajax()) {
-            $data = TingkatanKelas::latest()->get();
+            $data = TingkatanKelas::where('user_id', auth()->user()->id)->latest()->get();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
-                    $button = '<button type="button" id="'.$data->id.'" class="edit btn btn-mini btn-info shadow-sm">Edit</button>';
-                    $button .= '&nbsp;&nbsp;&nbsp;<button type="button" id="'.$data->id.'" class="delete btn btn-mini btn-danger shadow-sm">Delete</button>';
+                    $button = '<button type="button" id="' . $data->id . '" class="edit btn btn-mini btn-info shadow-sm">Edit</button>';
+                    $button .= '&nbsp;&nbsp;&nbsp;<button type="button" id="' . $data->id . '" class="delete btn btn-mini btn-danger shadow-sm">Delete</button>';
                     return $button;
                 })
                 ->rawColumns(['action'])
                 ->addIndexColumn()
                 ->make(true);
         }
-        
-        return view('admin.referensi.tingkatan-kelas');
+        return view('admin.referensi.tingkatan-kelas', ['mySekolah' => User::sekolah()]);
     }
 
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         // validasi
         $rules = [
             'tingkat'  => 'required|max:50',
@@ -47,6 +49,7 @@ class TingkatanKelasController extends Controller
         }
 
         $status = TingkatanKelas::create([
+            'user_id' => auth()->user()->id,
             'name'  => $request->input('tingkat'),
             'user_id' => $request->user()->id,
             'sekolah_id' => $request->user()->id_sekolah,
@@ -58,7 +61,8 @@ class TingkatanKelasController extends Controller
             ]);
     }
 
-    public function edit($id) {
+    public function edit($id)
+    {
         $tingkat = TingkatanKelas::find($id);
 
         return response()
@@ -67,7 +71,8 @@ class TingkatanKelasController extends Controller
             ]);
     }
 
-    public function update(Request $request) {
+    public function update(Request $request)
+    {
         // validasi
         $rules = [
             'tingkat'  => 'required|max:50',
@@ -85,8 +90,10 @@ class TingkatanKelasController extends Controller
                     'errors' => $validator->errors()->all()
                 ]);
         }
-
-        $status = TingkatanKelas::whereId($request->input('hidden_id'))->update([
+        $status = TingkatanKelas::where([
+            ['id', $request->input('hidden_id')],
+            ['user_id', auth()->user()->id]
+        ])->update([
             'name'  => $request->input('tingkat'),
             'user_id' => $request->user()->id,
             'sekolah_id' => $request->user()->id_sekolah,
@@ -98,8 +105,13 @@ class TingkatanKelasController extends Controller
             ]);
     }
 
-    public function destroy($id) {
+    public function destroy($id)
+    {
         $tingkat = TingkatanKelas::find($id);
+        if ($tingkat->user_id != auth()->user()->id) {
+            return;
+        }
+
         $tingkat->delete();
     }
 }
